requirements.txt
# Seg_diffusion: segmentation and diffusion for BraTS-style data
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.62.0
scipy>=1.7.0
scikit-learn>=1.0.0
diffusers>=0.29.0
transformers>=4.41.0
accelerate>=0.30.0
safetensors>=0.4.0
torchmetrics>=0.11.0
wandb>=0.15.0
pandas>=1.3.0
einops>=0.6.0
skimage>=0.19.0
